refactor(Arch): Arch ready to work
