+++ ui/src/layouts/storage/presentation/controller/mod.rs
@@ -1,2 +1,3 @@
 pub mod controller;
+pub mod coroutine;
 pub mod events;
